refactor: папка image вынесена в src
